backend/requirements.txt
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
django-filter==2.4.0
django_extensions==3.1.3
django-rest-knox==4.*
psycopg2>=2.8.6,<2.8.7
uWSGI>=2.0.19.1,<2.1
